+++ tasks/ExtractVersionFromTagTask/run.sh
@@ -1,3 +1,4 @@
-export INPUT_ProjectFolderPath="."
+export INPUT_ProjectFolderPath="./../../"
+export INPUT_TagPrefixMatch=""
 tsc &&
 node extract-version-from-tag.js