+++ tasks/UpdateiOSVersionInfoPlistTask/update-ios-version-info-plist.ts
@@ -29,7 +29,7 @@ async function run() {
             process.exit(1);
         }
 
-        var fileContent = fs.readFileSync(infoPlistPath, 'UTF8');
+        var fileContent = fs.readFileSync(infoPlistPath, { encoding: 'utf-8' });
         task.debug(JSON.stringify(fileContent));
 
         var obj = plist.parse(fileContent);